Moved hubbard to new file, better version check
